style: the lint gate over the whole tree, which the per-package runs had not seen
Seven files, all from this branch's own changes plus one the 386 job forced.

plugin.Unusable takes a type parameter instead of an any: the guard reads the
dynamic value through reflect either way, so the caller is unaffected, and the
signature stops claiming to accept anything in particular. Its test fixtures
gained the compile-time assertions that make a broken one a build error.

The filesystem-magic switch in the file writer's bench helper compares at
uint32. Statfs_t.Type is int64 on amd64 and int32 on 386, so neither width can
be written down — an int64 conversion reads as redundant on the host and is
required on the other. Every magic fits in 32 bits.

The rest is the house style the per-package runs had let through: an intent
comment on three blocks, a goroutine-lifetime paragraph on the three new
session tests, and a const moved above the functions it had drifted between.

+++ internal/service/writer/file/file_linux_bench_test.go
@@ -58,14 +58,14 @@ const openFlagsForRaw int = os.O_APPEND | os.O_CREATE | os.O_WRONLY | syscall.O_
 // anything else is reported as unrecognised rather than silently mislabelled.
 const (
 	// magicTmpfs is TMPFS_MAGIC — the page cache, with no device behind it.
-	magicTmpfs int64 = 0x01021994
+	magicTmpfs uint32 = 0x01021994
 	// magicExt is EXT2/3/4's shared magic; this machine's /home is ext4.
-	magicExt int64 = 0xEF53
+	magicExt uint32 = 0xEF53
 	// magicOverlay is OVERLAYFS_SUPER_MAGIC, which a container's writable
 	// layer is.
-	magicOverlay int64 = 0x794C7630
+	magicOverlay uint32 = 0x794C7630
 	// magicXFS is XFS_SUPER_MAGIC, common on cloud instances.
-	magicXFS int64 = 0x58465342
+	magicXFS uint32 = 0x58465342
 )
 
 // benchLine is the payload every row carries unless it says otherwise: the text
@@ -102,9 +102,10 @@ func fsName(b *testing.B, dir string) string {
 		//: an unreadable filesystem type is stated, never guessed.
 		return "unknown"
 	}
-	//: Statfs_t.Type is int64 on 64-bit Linux and int32 on 386, so the
-	//: comparison is made at one width rather than at the platform's.
-	switch int64(st.Type) {
+	//: Statfs_t.Type is int64 on 64-bit Linux and int32 on 386, so neither
+	//: width can be written down here. Every magic fits in 32 bits, so the
+	//: comparison is made there — a real conversion on both platforms.
+	switch uint32(st.Type) {
 	//: the page cache, no device behind it.
 	case magicTmpfs:
 		return "tmpfs"
@@ -120,7 +121,7 @@ func fsName(b *testing.B, dir string) string {
 	//: anything else is visibly unrecognised, and the magic goes to the log
 	//: rather than into a row name nobody can compare across machines.
 	default:
-		b.Logf("statfs %s: unrecognised magic %#x", dir, int64(st.Type))
+		b.Logf("statfs %s: unrecognised magic %#x", dir, uint32(st.Type))
 		return "unrecognised"
 	}
 }
